Fix request role dialog buttons missing with too many items.
The "don't ask again" check box was added as a custom view to the
alert dialog, and because we were using the list view in content panel
at the same time, AlertDialogLayout refuses to handle layout with its
special logic when both panels are used, and causes list view to take
up all the space with no space left for button bar. This can happen
for any device in landscape mode if there are a lot of candidate apps.

This CL uses a custom list view inside the custom view added to the
alert dialog, so that AlertController will hide the content
panel. This way AlertDialogLayout will only see the custom panel and
layout properly.

Fixes: 135549535
Test: manual && atest RoleManagerTest
Change-Id: I9e376279f10e8258f531ab27349c8cfb80d6f575